requirements.txt
python -m venv envfastapi
envfastapi\Scripts\activate

pip install fastapi
pip install uvicorn